putting mail prop into prop file
