+++ src/main/java/com/mendusa/transactions/repository/TransactionRepository.java
@@ -8,5 +8,4 @@ import java.util.List;
 
 @Repository
 public interface TransactionRepository extends JpaRepository<Transaction, Long> {
-    List<Transaction> findAllByOrderByIdDesc();// todo: not needed
 }
