+++ src/main/resources/application.properties
@@ -13,3 +13,8 @@ spring.mail.host=smtp.elasticemail.com
 spring.mail.username=nwajeigoddowell@gmail.com
 spring.mail.password=43AC4247B5352EA17D5D8C6095901084BEE5
 spring.mail.port=2525
+
+spring.mail.properties.transport.protocol=smtp
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
+spring.mail.properties.mail.debug=true
